working on the segue...
